Desc Product to Text

+++ app/Http/Controllers/Admin/ProductResourceController.php
@@ -51,7 +51,7 @@ class ProductResourceController extends Controller
         $request->validate([
             'name' => 'required',
             'price' => 'required|numeric',
-            'desc' => 'required|max:255',
+            'desc' => 'required',
             'stock' => 'required|numeric',
             'weight' => 'required|numeric',
             'categories' => 'required',
@@ -124,7 +124,7 @@ class ProductResourceController extends Controller
         $request->validate([
             'name' => 'required',
             'price' => 'required|numeric',
-            'desc' => 'required|max:255',
+            'desc' => 'required',
             'stock' => 'required|numeric',
             'weight' => 'required|numeric',
         ]);
@@ -151,29 +151,24 @@ class ProductResourceController extends Controller
      */
     public function destroy($id)
     {
+        $cart = DB::table('carts')->where('product_id', $id)->get()->count();
 
-    $cart = DB::table('carts')->whereNull('product_id', $id);
-    $cart1 = DB::table('carts')->where('product_id', $id);
-
-    if($cart1){
-        return redirect('admin/product')->with('danger','The product is still in the customers cart');
-      }
-
-    if ($cart){
-        $image = DB::table('product_images')->where('product_id', $id);
-        $image->delete();
-  
-        $category = DB::table('product_category_details')->where('product_id', $id);
-        $category->delete();
-
-        $discount = DB::table('discounts')->where('product_id', $id);
-        $discount->delete();
-        
-        Product::where('id',$id)->delete();
-        return redirect('admin/product')->with('success','Product has been deleted');
-      }
-
-     
+        if($cart && $cart != null && $cart != 0){
+            return redirect('admin/product')->with('danger','Product are on user cart');
+        }
+        else{
+            $image = DB::table('product_images')->where('product_id', $id);
+            $image->delete();
+      
+            $category = DB::table('product_category_details')->where('product_id', $id);
+            $category->delete();
+    
+            $discount = DB::table('discounts')->where('product_id', $id);
+            $discount->delete();
+            
+            Product::where('id',$id)->delete();
+            return redirect('admin/product')->with('success','Product has been deleted');
+        }
     }
 
     public function uploadImage(Request $request)
@@ -202,8 +197,6 @@ class ProductResourceController extends Controller
     public function destroyCategory(Product $product, Request $request)
     {
         $product->categories()->detach($request->id);
-        // $category = ProductCategoryDetail::find($id);
-        // $category->delete();
         return redirect()->back()->with('success','Category Product has been deleted');
     }
